refaector: aplicação das novas Exceptions

+++ ProjetoDevTrail.Application/UseCase/Clients/DeleteClient/DeleteClientHandler.cs
@@ -1,4 +1,5 @@
-﻿using ProjetoDevTrail.Infra.Repositories.ClientRepositories;
+﻿using ProjetoDevTrail.Application.Utils.Exceptions;
+using ProjetoDevTrail.Infra.Repositories.ClientRepositories;
 
 namespace ProjetoDevTrail.Application.UseCase.Clients.DeleteClient
 {
@@ -6,6 +7,9 @@ namespace ProjetoDevTrail.Application.UseCase.Clients.DeleteClient
     {
         public async Task HandleAsync(Guid id)
         {
+            var clienteExists = await repo.GetByIdAsync(id);
+            if (clienteExists == null)
+                throw new NotFoundException("Cliente não encontrado");
             await repo.DeleteAsync(id);
         }
     }
